Move date-utils fn
